ATM Bank Account
This code represents a basic ATM system. It prompts the user for a PIN input, and if the PIN is correct, it asks the user to choose an operation, such as withdrawal or balance check. If the user selects withdrawal, the code prompts for an amount input, and if the available balance is greater than or equal to the withdrawal amount, the transaction is completed, and the available balance is updated. If the balance is insufficient, the user is notified with a "Your balance is lower" message. If the user chooses to check the balance, the current available balance is displayed. If the user enters an incorrect PIN, an "Incorrect Pin Code!!" message is displayed.